requirements.txt
torch
torchvision
numpy
opencv-python
scipy
pyyaml
tqdm
huggingface_hub
ultralytics
smplx

